Create status.conf
This goes in /etc/apache2/mods-enabled/status.conf 
I commented out the "Require local" line, it should be replaced with a line similar to "Require ip 192.0.2.0/24" with the subnet that Metricbeat runs in.  Commenting out Require local and not adding a Require ip line makes the status page available to the outside world, and this is not necessary.